+++ package.json
@@ -20,35 +20,35 @@
     "@angular/platform-browser": "8.0.0",
     "@angular/platform-browser-dynamic": "8.0.0",
     "@angular/router": "8.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^4.2.1",
+    "@ng-bootstrap/ng-bootstrap": "^4.2.2",
     "bootstrap": "^3.4.1",
-    "core-js": "^2.6.9",
-    "rxjs": "^6.5.2",
-    "rxjs-compat": "^6.5.2",
+    "core-js": "^2.6.11",
+    "rxjs": "^6.5.4",
+    "rxjs-compat": "^6.5.4",
     "tough-cookie": "^2.5.0",
-    "tslib": "^1.9.0",
+    "tslib": "^1.11.1",
     "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.0",
-    "@angular/cli": "^8.0.1",
-    "@angular/compiler-cli": "^8.0.0",
+    "@angular-devkit/build-angular": "^0.900.5",
+    "@angular/cli": "^8.3.25",
+    "@angular/compiler-cli": "^8.2.14",
     "@angular/language-service": "8.0.0",
     "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "^2.0.6",
-    "@types/node": "^6.14.6",
-    "codelyzer": "^5.0.1",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^6.14.9",
+    "codelyzer": "^5.2.1",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
-    "karma": "^4.1.0",
+    "karma": "^4.4.1",
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.4.3",
     "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.4.2",
+    "protractor": "^5.4.3",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "^3.4.5"
+    "typescript": "^3.8.3"
   }
 }
