+++ src/app/audio-experiments/audio-experiments.component.html
@@ -1,4 +1,4 @@
-<div class=''>
+<div class='' (touchstart)="swipe($event, 'start')" (touchend)="swipe($event, 'end')">
   <p>audio-experiments works!</p>
   <h2>{{swipe}}</h2>
 </div>
